DefaultFile.py was altered to correctly fit the PEP-8 standard that we are using currently.
EntityGenerator.py then had a function added to detect input from the user to allow them to save the currently photo given to the function whenever they pressed 'p'.

I also started creating the parent class of Shape as well as creating the child classes of Rectangle and Circle. This approach allows the user enter values into each class and for it to draw on the screen. Using inheritance this will allow me to store a list of Shapes where I can then call the function of draw_shape() to then draw them all at once even if their pygame.draw syntax is different and requires different parameters.

I've also considered allowing the user to click and draw on the screen, but this doesn't quite full-fill the requirements set by the contract.